+++ cppcache/integration-test/testSerialization.cpp
@@ -51,7 +51,7 @@ std::shared_ptr<T> duplicate(const std::shared_ptr<T>& orig) {
   auto dout = getHelper()->getCache()->createDataOutput();
   dout->writeObject(orig);
 
-  uint32_t length = 0;
+  size_t length = 0;
   const uint8_t* buffer = dout->getBuffer(&length);
   auto din = getHelper()->getCache()->createDataInput(buffer, length);
   din->readObject(result);
